Add files via upload
## Case study 02 - Project 01

Last year, a local college implemented rooftop gardens as a way to promote energy efficiency and save money. Write a program that will allow the user to enter the energy bills from January to December for the year prior to going green. Next, allow the user to enter the energy bills from January to December of the past year after going green. **The program should calculate the energy difference from the two years and display the two years worth of data, along with the savings**. Additionally:

- The savings array should be printed to a file called savings.txt. 
- Functions must be used to perform the required tasks. 
- The program should validate the user inputs so only valid data are accepted for the calculation process.
- The program should display the two months in which the maximum and minimum amount of savings have occurred.
- The program should return the amount of saving of the month entered by the user.